+++ packages/view-types/heatmap/src/HeatmapOptions.js
@@ -1,6 +1,6 @@
 import React, { useCallback } from 'react';
 import { debounce } from 'lodash-es';
-import { Slider, TableCell, TableRow } from '@material-ui/core';
+import { Checkbox, Slider, TableCell, TableRow } from '@material-ui/core';
 import { usePlotOptionsStyles, OptionsContainer, OptionSelect } from '@vitessce/vit-s';
 import { GLSL_COLORMAPS } from '@vitessce/gl';
 
@@ -10,12 +10,18 @@ export default function HeatmapOptions(props) {
     setGeneExpressionColormap,
     geneExpressionColormapRange,
     setGeneExpressionColormapRange,
+    tooltipsVisible,
+    setTooltipsVisible,
   } = props;
 
   const classes = usePlotOptionsStyles();
 
   function handleGeneExpressionColormapChange(event) {
     setGeneExpressionColormap(event.target.value);
+  }
+
+  function handleTooltipsVisibilityChange(event) {
+    setTooltipsVisible(event.target.checked);
   }
 
   function handleColormapRangeChange(event, value) {
@@ -45,6 +51,25 @@ export default function HeatmapOptions(props) {
               <option key={cmap} value={cmap}>{cmap}</option>
             ))}
           </OptionSelect>
+        </TableCell>
+      </TableRow>
+      <TableRow>
+        <TableCell className={classes.labelCell}>
+          Tooltips Visible
+        </TableCell>
+        <TableCell className={classes.inputCell}>
+          <Checkbox
+            className={classes.checkbox}
+              /**
+               * We have to use "checked" here, not "value".
+               * The checkbox state is not persisting with value.
+               * For reference, https://v4.mui.com/api/checkbox/
+               */
+            checked={tooltipsVisible}
+            onChange={handleTooltipsVisibilityChange}
+            name="gene-expression-colormap-option-toltip-visibility"
+            color="default"
+          />
         </TableCell>
       </TableRow>
       <TableRow>