+++ packages/view-types/scatterplot/src/ScatterplotOptions.js
@@ -21,6 +21,8 @@ export default function ScatterplotOptions(props) {
     setCellOpacityMode,
     cellSetLabelsVisible,
     setCellSetLabelsVisible,
+    tooltipsVisible,
+    setTooltipsVisible,
     cellSetLabelSize,
     setCellSetLabelSize,
     cellSetPolygonsVisible,
@@ -55,6 +57,10 @@ export default function ScatterplotOptions(props) {
 
   function handleLabelVisibilityChange(event) {
     setCellSetLabelsVisible(event.target.checked);
+  }
+
+  function handleTooltipsVisibilityChange(event) {
+    setTooltipsVisible(event.target.checked);
   }
 
   function handleLabelSizeChange(event, value) {
@@ -95,6 +101,25 @@ export default function ScatterplotOptions(props) {
             checked={cellSetLabelsVisible}
             onChange={handleLabelVisibilityChange}
             name="scatterplot-option-cell-set-labels"
+            color="default"
+          />
+        </TableCell>
+      </TableRow>
+      <TableRow>
+        <TableCell className={classes.labelCell}>
+          Tooltips Visible
+        </TableCell>
+        <TableCell className={classes.inputCell}>
+          <Checkbox
+            className={classes.checkbox}
+              /**
+               * We have to use "checked" here, not "value".
+               * The checkbox state is not persisting with value.
+               * For reference, https://v4.mui.com/api/checkbox/
+               */
+            checked={tooltipsVisible}
+            onChange={handleTooltipsVisibilityChange}
+            name="scatterplot-option-toltip-visibility"
             color="default"
           />
         </TableCell>